+++ http-request/guestCart.http
@@ -14,7 +14,7 @@ Content-Type: application/json
 Accept: application/json
 
 {
-  "bookId": 3,
+  "bookId": 2,
   "quantity": 3
 }
