+++ src/main/java/com/dooray/bookstorecarts/service/GuestCartService.java
@@ -1,28 +1,23 @@
 package com.dooray.bookstorecarts.service;
 
-import com.dooray.bookstorecarts.exception.CartAlreadyExistsException;
 import com.dooray.bookstorecarts.exception.CartNotFoundException;
 import com.dooray.bookstorecarts.redisdto.GuestCart;
-import com.dooray.bookstorecarts.repository.GuestCartRepository;
 import com.dooray.bookstorecarts.response.GuestCartResponse;
-import lombok.RequiredArgsConstructor;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-
 @Service
-@RequiredArgsConstructor
 public class GuestCartService {
-    private final GuestCartRepository guestCartRepository;
-
-    public GuestCartResponse getCartBySessionId(String sessionId) {   // 세션 아이디로 비회원카트 반환
-        GuestCart guestCart = guestCartRepository.findBySessionId(sessionId)
-                .orElseThrow(() -> new CartNotFoundException(sessionId));
+    public GuestCartResponse getCartBySession(HttpSession session) {
+        GuestCart guestCart = (GuestCart) session.getAttribute("guestCart");
+        if (guestCart == null) {
+            throw new CartNotFoundException(session.getId());
+        }
 
         return new GuestCartResponse(guestCart);
     }
 
-    public void deleteGuestCart(String sessionId) {
-        guestCartRepository.delete(sessionId);
+    public void deleteGuestCart(HttpSession session) {
+        session.invalidate();
     }
 }
